Completed schema for a room Seperated code that used to live in seed.js into it's own model folder
